Removed unused definitions.
The spec defined 'Additional Window Controls' and 'Isolated Web App',
presumably for use with borderless.

Neither of these terms were referenced. The former is just a thing users
can build, not a formal concept. The latter is supposed to be defined in
a separate spec, so shouldn't have a definition here.