Fix the logic of whether the app has a new tab button.
Now has an explicit concept of "has a new tab button" (analogous to "has
a home tab"). The parser algorithm no longer (accidentally) sets the new
tab URL to the start URL if it is within home tab scope.

Add a note clarifying why the new tab button is not allowed to be in
home tab scope.

Also note the start URL default, which is easy to miss as it is buried
in the parser algorithm.